Separated Euler2D into its own file and Solver into the base file
